initial work on chat

+++ src/shared/MessageUtil.c
@@ -13,7 +13,8 @@ const int n = 221;
 
 /* Encrypts a given message and places it in the provided buffer */
 /* Returns -1 if the size of msg and buffer do not match */
-int encryptMessage(int* buffer, char* msg, size_t bufferSize) {
+int encryptMessage(long unsigned int *buffer, char* msg, size_t bufferSize) 
+{
     if (bufferSize < strlen(msg) + 1)
     { // Include space for the null terminator
         return -1;
